change folder layout
