+++ . prettierignore
@@ -1,0 +1,2 @@
+# package.json is formatted by package managers, so we ignore it here
+package.json
