+++ .cspell.json
@@ -1,0 +1,49 @@
+{
+    "version": "0.1",
+    "$schema": "https://raw.githubusercontent.com/streetsidesoftware/cspell/master/cspell.schema.json",
+    "language": "en",
+    "words": [
+        "bitauth",
+        "bitjson",
+        "blocksmc",
+        "cimg",
+        "circleci",
+        "codecov",
+        "commitlint",
+        "dependabot",
+        "editorconfig",
+        "esnext",
+        "execa",
+        "exponentiate",
+        "funcraft",
+        "globby",
+        "gommehd",
+        "hivemc",
+        "leaderboard",
+        "leaderboards",
+        "libauth",
+        "manacube",
+        "mccentral",
+        "mineplex",
+        "minesaga",
+        "mkdir",
+        "munchymc",
+        "prettierignore",
+        "sandboxed",
+        "timonlia",
+        "transpiled",
+        "typedoc",
+        "universocraft",
+        "untracked",
+        "veltpvp",
+        "wynncraft"
+    ],
+    "flagWords": [],
+    "ignorePaths": [
+        "package.json",
+        "package-lock.json",
+        "yarn.lock",
+        "tsconfig.json",
+        "node_modules/**"
+    ]
+}
