+++ .github/CONTRIBUTING.md
@@ -1,0 +1,3 @@
+# Example Contributing Guidelines
+
+This is an example of GitHub's contributing guidelines file. Check out GitHub's [CONTRIBUTING.md help center article](https://help.github.com/articles/setting-guidelines-for-repository-contributors/) for more information.
