+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,7 @@
+- **What kind of change does this PR introduce?** (Bug fix, feature, docs update, ...)
+
+- **What is the current behavior?** (You can also link to an open issue here)
+
+- **What is the new behavior (if this is a feature change)?**
+
+- **Other information**:
