+++ .gitignore
@@ -1,3 +1,10 @@
+
+diff
+.idea/*
+.nyc_output
+build
 node_modules
-.eslintrc.json
-.DS_Store
+test
+src/**.js
+coverage
+*.log
