+++ .vscode/extensions.json
@@ -1,0 +1,9 @@
+
+{
+    "recommendations": [
+      "dbaeumer.vscode-eslint",
+      "esbenp.prettier-vscode",
+      "eamodio.gitlens",
+      "streetsidesoftware.code-spell-checker",
+    ]
+  }
