+++ .vscode/settings.json
@@ -1,0 +1,7 @@
+{
+    "cSpell.userWords": [], // only use words from .cspell.json
+    "cSpell.enabled": true,
+    "editor.formatOnSave": true,
+    "typescript.tsdk": "node_modules/typescript/lib",
+    "typescript.enablePromptUseWorkspaceTsdk": true
+  }
