+++ src/index.ts
@@ -1,1 +1,13 @@
-module.exports = require("./util.js");
+export * from './lib/blocksmc'
+export * from './lib/funcraft'
+export * from './lib/gommehd'
+export * from './lib/hivemc'
+export * from './lib/manacube'
+export * from './lib/mccentral'
+export * from './lib/mineplex'
+export * from './lib/minesaga'
+export * from './lib/munchymc'
+export * from './lib/timonlia'
+export * from './lib/universocraft'
+export * from './lib/veltpvp'
+export * from './lib/wynncraft'
